+++ src/main/scala/org/codeprose/api/ProjectSummary.scala
@@ -1,0 +1,6 @@
+package org.codeprose.api
+
+import org.codeprose.util.DynamicPropertyMap
+
+
+class ProjectSummary() extends DynamicPropertyMap
