+++ src/main/scala/org/codeprose/api/common/TokenProperties.scala
@@ -1,0 +1,5 @@
+package org.codeprose.api.common
+
+class TokenProperties {
+  
+}
