+++ src/main/scala/org/codeprose/api/scalalang/SummaryProperties.scala
@@ -1,0 +1,5 @@
+package org.codeprose.api.scalalang
+
+class SummaryProperties {
+  
+}
