+++ src/main/scala/org/codeprose/api/scalalang/TokenProperties.scala
@@ -1,0 +1,5 @@
+package org.codeprose.api.scalalang
+
+class TokenProperties {
+  
+}
