+++ src/main/scala/org/codeprose/consumer/WriterLatex.scala
@@ -1,7 +1,9 @@
 package org.codeprose.consumer
-//
-//class WriterLatex extends Consumer {
-//  def generateOutput(): Unit = {
-//    ???
-//  }
-//}
+
+import org.codeprose.api.ProjectInfo
+
+class WriterLatex extends Consumer {
+  def generateOutput(projectInfo: ProjectInfo) : Unit = {
+    ???
+  }
+}
